move decode_single_frame_from_compressed_data() to Decoder
